[Issue #25] Added namespace for Pixmicat\Pio.
